Finish command queue
We now have a single-producer single-consumer command queue that is
lock-free on both ends (which is fun).
If I do need a case where we have more than one producer, I'll lock the
producing end of the queue, but the consumer will remain lock-free
(because there is only one consumer in any case). This would address the
case where we have, for example, an I/O task, a logic task and a comm
task, where both the logic task and the comm task can create CROBs.