Implement alternative way to handle mappings in the Consumer
The default way remains as-is.
The new way sorts the mappings so finding the mapping to or from a point
is O(lg(n)). This does mean that constructing the mapping requires
sorting, and is O(n lg(n)); and there's an overhead of two pointers per
mapping (so 400 pointers in the default configuration of 200 points).
The overhead is why the default mapping (which is compact but O(n) in both
construction and search) remains default.