Add metadata to points
This adds flags and a time stamp to each point in the data store.
It also moves the Timestamp class to the Core namespace, and adds Flag
definitions according to what you'd normally find in a SCADA protocol
(such as DNP3).

The overhead added to the memory footprint is fairly small: 56 bits per
permitted concurrent transition (+1) per point if packed. I think the
benefits outweight the costs enough to not make this optional.

Signed-off-by: Ronald Landheer-Cieslak <rlc@vlinder.ca>